+++ src/api/domain/tmdb/popular_movie.go
@@ -28,7 +28,7 @@ package tmdb
 type PopularMovieRequest struct {
 	ApiKey   string `json:"apiKey"`
 	Language string `json:"language"`
-	Page     int64  `json:"page"`
+	Page     int    `json:"page"`
 }
 
 type PopularMovieResponse struct {
